refactor(search): move shared contract into workspace
