+++ services/search-shared/src/search-shared.ts
