+++ services/search-worker/src/search-results.ts
@@ -1,4 +1,4 @@
-import { type SearchResult } from './search-shared'
+import { type SearchResult } from 'search-shared'
 
 type RetrievedMatch = {
 	rawId: string
