--- services/search-worker/src/search-shared.ts
@@ -1,1 +1,0 @@
-export * from '../../../other/search/search-service.ts'
